refactor: migrate committees page to feature architecture
- Move CommitteeModal into src/features/committees/components/
- Add CommitteesTable component
- Add CommitteesContainer with AppPageLayout + FloatingActions
- Slim down CommitteesPageClient to delegate to CommitteesContainer
