+++ src/features/committees/index.ts
@@ -1,0 +1,1 @@
+export {CommitteesContainer} from './containers/CommitteesContainer';
